requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
pydantic==1.10.2
PyYAML==6.0
typing_extensions==4.3.0
